# Improve performance of Gorilla float block decoding